+++ HFS/changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.11.0] - unreleased
 ### Changed
+- Massive improvement in writing speed using buffered stream
 - Improved MNG decoding performance
 - .NET 10